# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wealthboard API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
paths:
...
  /goals:
    get:
      operationId: listGoals
      summary: List the current user's goals
      description: Authenticate with a browser session or a Bearer API key that grants `portfolio:read`.
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Owner-scoped goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createGoal
      summary: Create a financial goal and contribution plan
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`. Monetary inputs are major-unit decimal strings. Idempotency may be supplied in the body or `Idempotency-Key`; when both are present they must match.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/MutationCSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalMutationRequest"
      responses:
        "200":
          description: An idempotent goal creation was replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalMutationResult"
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalMutationResult"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /goals/{id}:
    get:
      operationId: getGoal
      summary: Return one goal owned by the current user
      description: Authenticate with a browser session or a Bearer API key that grants `portfolio:read`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Owner-scoped goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    put:
      operationId: updateGoal
      summary: Replace a financial goal and contribution plan
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`. Any body idempotency key is ignored.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/MutationCSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalMutationRequest"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationStatusResponse"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteGoal
      summary: Delete a financial goal
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/MutationCSRFToken"
      responses:
        "204": { description: Goal deleted }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /goals/{id}/status:
    patch:
      operationId: setGoalStatus
      summary: Set a financial goal status
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/MutationCSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalStatusRequest"
      responses:
        "200":
          description: Goal status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationStatusResponse"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /goals/{id}/milestones:
    get:
      operationId: listGoalMilestones
      summary: List milestones for one goal
      description: Authenticate with a browser session or a Bearer API key that grants `portfolio:read`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Owner-scoped goal milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalMilestone"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createGoalMilestone
      summary: Create a milestone for a financial goal
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`. The target amount is a major-unit decimal string in the goal currency.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/MutationCSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalMilestoneCreateRequest"
      responses:
        "201":
          description: Goal milestone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceIDResponse"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /goals/{id}/milestones/{milestoneId}:
    delete:
      operationId: deleteGoalMilestone
      summary: Delete a milestone from a financial goal
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - name: milestoneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/MutationCSRFToken"
      responses:
        "204": { description: Goal milestone deleted }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /goals/{id}/alerts/dismiss:
    post:
      operationId: dismissGoalAlert
      summary: Dismiss the current monthly alert for a goal
      description: Requires `portfolio:write`. Browser sessions also require a trusted Origin and `X-CSRF-Token`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/MutationCSRFToken"
      responses:
        "200":
          description: Goal alert dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationStatusResponse"
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wealthboard_session
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: wbk_v1
      description: API-key scopes are enforced per operation as documented in each operation description.
  parameters:
...
    MutationCSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when authenticating with `sessionCookie`; ignored for Bearer API keys.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: UUID request key. When also supplied in the JSON body, both values must match.
      schema:
        type: string
        format: uuid
...
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    MinorUnits:
      type: string
      description: Integer monetary amount in the currency's minor units.
      pattern: "^-?[0-9]+$"
    DecimalString:
      type: string
      description: Exact decimal value serialized as a string.
      pattern: "^-?[0-9]+(?:\\.[0-9]+)?$"
    NonNegativeDecimalString:
      type: string
      description: Exact non-negative decimal value serialized as a string.
      pattern: "^[0-9]+(?:\\.[0-9]+)?$"
    PositiveDecimalString:
      type: string
      description: Exact positive decimal value serialized as a string.
      pattern: "^(?:0*[1-9][0-9]*)(?:\\.[0-9]+)?$|^0*\\.[0-9]*[1-9][0-9]*$"
    CurrencyCode:
      type: string
      pattern: "^[A-Z]{3}$"
...
    ResourceIDResponse:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id: { type: string, format: uuid }
    MutationStatusResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string }
...
    GoalMutationRequest:
      type: object
      additionalProperties: false
      required:
        [
          name,
          targetAmount,
          currency,
          targetDate,
          assumedAnnualReturn,
          plannedContribution,
          planStartDate,
        ]
      properties:
        idempotencyKey: { type: [string, "null"], format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: [string, "null"], maxLength: 500 }
        targetAmount:
          $ref: "#/components/schemas/PositiveDecimalString"
        currentAmount:
          $ref: "#/components/schemas/NonNegativeDecimalString"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        targetDate: { type: string, format: date }
        linkedAccountId: { type: [string, "null"], format: uuid }
        icon: { type: string, maxLength: 50, default: Target }
        status:
          {
            type: string,
            enum: [active, paused, completed, cancelled],
            default: active,
          }
        priority: { type: integer, minimum: 0, maximum: 100 }
        assumedAnnualReturn: { type: number, minimum: 0, maximum: 100 }
        plannedContribution:
          $ref: "#/components/schemas/NonNegativeDecimalString"
        frequency:
          {
            type: string,
            enum: [weekly, monthly, quarterly, annually, custom],
            default: monthly,
          }
        planStartDate: { type: string, format: date }
        planEndDate: { type: string, format: date }
    GoalMutationResult:
      type: object
      additionalProperties: false
      required: [id, replayed]
      properties:
        id: { type: string, format: uuid }
        replayed: { type: boolean }
    GoalStatusRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, enum: [active, paused, completed, cancelled] }
    GoalMilestoneCreateRequest:
      type: object
      additionalProperties: false
      required: [name, targetAmount]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        targetAmount:
          $ref: "#/components/schemas/PositiveDecimalString"
        targetDate: { type: string, format: date }
...
    GoalLinkedAccount:
      type: object
      additionalProperties: false
      required: [id, name, currency]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        currency: { type: string }
    GoalPlan:
      type: object
      additionalProperties: false
      required: [plannedContributionMinor, frequency, startDate, endDate]
      properties:
        plannedContributionMinor:
          $ref: "#/components/schemas/MinorUnits"
        frequency: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: [string, "null"], format: date }
    Goal:
      type: object
      additionalProperties: false
      required:
        [
          id,
          name,
          description,
          targetAmountMinor,
          currentAmountMinor,
          currentAmountCurrency,
          currency,
          targetDate,
          linkedAccount,
          icon,
          status,
          priority,
          assumedAnnualReturnBps,
          progressPercent,
          trackingStatus,
          valueIncomplete,
          missingCurrencies,
          plan,
          projection,
        ]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, "null"] }
        targetAmountMinor:
          $ref: "#/components/schemas/MinorUnits"
        currentAmountMinor:
          $ref: "#/components/schemas/MinorUnits"
        currentAmountCurrency: { type: string }
        currency: { type: string }
        targetDate: { type: string, format: date }
        linkedAccount:
          oneOf:
            - $ref: "#/components/schemas/GoalLinkedAccount"
            - type: "null"
        icon: { type: string }
        status: { type: string }
        priority: { type: integer, format: int32 }
        assumedAnnualReturnBps: { type: integer, format: int32 }
        progressPercent:
          $ref: "#/components/schemas/DecimalString"
        trackingStatus:
          type: string
          enum: [on_track, behind, incomplete]
        valueIncomplete: { type: boolean }
        missingCurrencies:
          type: array
          items: { type: string }
        plan:
          oneOf:
            - $ref: "#/components/schemas/GoalPlan"
            - type: "null"
        projection:
          type: array
          items:
            $ref: "#/components/schemas/GoalProjectionPoint"
        scenarios:
          oneOf:
            - $ref: "#/components/schemas/GoalScenarios"
            - type: "null"
    GoalScenarios:
      type: object
      additionalProperties: false
      required: [savedPlan, requiredPace, lowerReturn]
      properties:
        savedPlan: { $ref: "#/components/schemas/GoalScenario" }
        requiredPace: { $ref: "#/components/schemas/GoalScenario" }
        lowerReturn: { $ref: "#/components/schemas/GoalScenario" }
    GoalScenario:
      type: object
      additionalProperties: false
      required:
        [
          monthlyContributionMinor,
          annualReturnBps,
          projectedAtTargetMinor,
          projectedProgressPercent,
          newContributionsMinor,
          estimatedGrowthMinor,
          estimatedCompletion,
          reachesTarget,
        ]
      properties:
        monthlyContributionMinor: { $ref: "#/components/schemas/MinorUnits" }
        annualReturnBps: { type: integer, format: int32 }
        projectedAtTargetMinor: { $ref: "#/components/schemas/MinorUnits" }
        projectedProgressPercent: { $ref: "#/components/schemas/DecimalString" }
        newContributionsMinor: { $ref: "#/components/schemas/MinorUnits" }
        estimatedGrowthMinor: { $ref: "#/components/schemas/MinorUnits" }
        estimatedCompletion: { type: [string, "null"], format: date }
        reachesTarget: { type: boolean }
    GoalProjectionPoint:
      type: object
      additionalProperties: false
      required: [date, projectedMinor, contributionsMinor, targetMinor]
      properties:
        date: { type: string, format: date-time }
        projectedMinor:
          $ref: "#/components/schemas/MinorUnits"
        contributionsMinor:
          $ref: "#/components/schemas/MinorUnits"
        targetMinor:
          $ref: "#/components/schemas/MinorUnits"
    GoalMilestone:
      type: object
      additionalProperties: false
      required:
        [
          id,
          goalId,
          name,
          targetAmountMinor,
          targetDate,
          status,
          progressPercent,
          remainingMinor,
        ]
      properties:
        id: { type: string, format: uuid }
        goalId: { type: string, format: uuid }
        name: { type: string }
        targetAmountMinor:
          $ref: "#/components/schemas/MinorUnits"
        targetDate: { type: [string, "null"], format: date }
        status: { type: string }
        progressPercent:
          $ref: "#/components/schemas/DecimalString"
        remainingMinor:
          oneOf:
            - $ref: "#/components/schemas/MinorUnits"
            - type: "null"
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
  responses:
...
    Problem:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"